+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Changed "Hiccup" effect to be more volatile if "Increase Knockback" is active.
+
 ## [2.0.7] - 2023-11-30
 
 - Changed "Fling Upwards" effect to force the player to stop grappling.